Move the spotbugs methods onto the separate file
